reorg code, move constants and more rocksdb-specifics to storage::rocksdb
